+++ src/global_styling/index.ts
@@ -7,4 +7,5 @@
  */
 
 export * from './reset/global_styles';
+export * from './variables/_shadow';
 export * from './mixins';
