+++ src/global_styling/mixins/index.ts
@@ -7,4 +7,3 @@
  */
 
 export * from './_helpers';
-export * from './_shadow';
