+++ src/themes/amsterdam/global_styling/mixins/_shadow.ts
@@ -9,7 +9,7 @@
 import { useEuiTheme, UseEuiTheme } from '../../../../services/theme';
 import { getShadowColor } from '../functions';
 import { createStyleHookFromMixin } from '../../../../global_styling/utils';
-import { _EuiShadowSizes } from '../../../../global_styling/mixins/_shadow';
+import { _EuiShadowSizes } from '../../../../global_styling/variables/_shadow';
 
 /**
  * Very subtle shadow used on small components, ya heard the type!
